server/requirements.txt
# Web Framework
PyYAML>=6.0
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# LangChain (PINNED – this is critical)
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.43
langchain-google-genai==0.0.6
langchain-groq==0.1.5

# Vector DB (legacy, Windows-safe)
pinecone==2.2.4



# PDF Parsing
pypdf==4.2.0

# Env
python-dotenv==1.0.1

# Utilities
requests==2.32.3
tqdm==4.67.1
pydantic==2.7.4
loguru==0.7.2
